Simplify API hooks
